Fedora 31 out-of-box README update
On Fedora 31, I had to install _both_ the X11 and Wayland packages listed in the 'requirements' section before I could successfully run the installation. To be clear, that was true regardless of whether I installed with or without the wayland tag. Updates the README to include this (admittedly rudimentary) suggestion.

I don't mind at all if this isn't merged, I just wanted to try to share the 'button presses that worked for me' update.