Change `WaveFormat` and `WaveCrop` signatures to accept a pointer to `Wave`
The raylib functions `WaveFormat` and `WaveCrop` accept a pointer to
`Wave` because the wave is updated in-place. However, the current Go
bindings pass a copy of the wave to the C functions, so it is not
updated. This commit changes the signatures of the Go API so they also
accept a pointer to `Wave`.